US124223: move list-item related files to list folder.
