+++ components/activity/editor/collection/custom/quiz/d2l-activity-collection-item-quiz.js
@@ -1,5 +1,5 @@
-import './d2l-activity-list-item-question.js';
-import './d2l-activity-list-item-section.js';
+import '../../../../list/custom/quiz/d2l-activity-list-item-question.js';
+import '../../../../list/custom/quiz/d2l-activity-list-item-section.js';
 
 import { css, LitElement } from 'lit-element/lit-element.js';
 import { HypermediaStateMixin, observableTypes } from '@brightspace-hmc/foundation-engine/framework/lit/HypermediaStateMixin.js';
