+++ components/activity/list/custom/quiz/d2l-activity-list-item-section.js
