+++ components/activity/list/custom/quiz/lang/de.js
