+++ components/activity/list/custom/quiz/lang/en.js
