+++ components/activity/list/custom/quiz/lang/es-es.js
