+++ components/activity/list/custom/quiz/lang/es.js
