+++ components/activity/list/custom/quiz/lang/fr-fr.js
