+++ components/activity/list/custom/quiz/lang/fr.js
