+++ components/activity/list/custom/quiz/lang/ja.js
