+++ components/activity/list/custom/quiz/lang/localization.js
