+++ components/activity/list/custom/quiz/lang/pt.js
